feat: centralize multer config
